chore(alb): Add types
