+++ internal/provider/alb/pool_schema.go
@@ -23,50 +23,6 @@ import (
 	fstringvalidator "github.com/FrangipaneTeam/terraform-plugin-framework-validators/stringvalidator"
 )
 
-type albPoolModel struct {
-	ID                       types.String `tfsdk:"id"`
-	EdgeGatewayID            types.String `tfsdk:"edge_gateway_id"`
-	EdgeGatewayName          types.String `tfsdk:"edge_gateway_name"`
-	Name                     types.String `tfsdk:"name"`
-	Enabled                  types.Bool   `tfsdk:"enabled"`
-	Description              types.String `tfsdk:"description"`
-	Algorithm                types.String `tfsdk:"algorithm"`
-	DefaultPort              types.Int64  `tfsdk:"default_port"`
-	GracefulTimeoutPeriod    types.Int64  `tfsdk:"graceful_timeout_period"`
-	Members                  types.Set    `tfsdk:"members"`
-	HealthMonitors           types.Set    `tfsdk:"health_monitors"`
-	PersistenceProfile       types.Object `tfsdk:"persistence_profile"`
-	PassiveMonitoringEnabled types.Bool   `tfsdk:"passive_monitoring_enabled"`
-
-	// CACertificateIDs         types.Set    `tfsdk:"ca_certificate_ids"`
-	// CNCheckEnabled           types.Bool   `tfsdk:"cn_check_enabled"`
-	// DomainNames              types.Set    `tfsdk:"domain_names"`
-}
-
-type member struct {
-	Enabled   types.Bool   `tfsdk:"enabled"`
-	IPAddress types.String `tfsdk:"ip_address"`
-	Port      types.Int64  `tfsdk:"port"`
-	Ratio     types.Int64  `tfsdk:"ratio"`
-}
-
-var memberAttrTypes = map[string]attr.Type{
-	"enabled":    types.BoolType,
-	"ip_address": types.StringType,
-	"port":       types.Int64Type,
-	"ratio":      types.Int64Type,
-}
-
-type persistenceProfile struct {
-	Type  types.String `tfsdk:"type"`
-	Value types.String `tfsdk:"value"`
-}
-
-var persistenceProfileAttrTypes = map[string]attr.Type{
-	"type":  types.StringType,
-	"value": types.StringType,
-}
-
 /*
 albPoolSchema
